+++ scripts/main.js
@@ -21,11 +21,11 @@ jQuery(document).ready(function($) {
 	 * Setup pause button
 	 */
 	btn_pause.on('click',function(){
-		$('#stage').okaynowthis('pause',function(){
+		$('#stage').okayNowThis('pause',function(){
 			btn_resume.removeAttr('disabled');
 			btn_pause.attr('disabled','disabled');
 			log_c.prepend('paused <br>');
-			log_z.text('.okaynowthis-paused');
+			log_z.text('.okayNowThis-paused');
 		});
 	});
 
@@ -33,11 +33,11 @@ jQuery(document).ready(function($) {
 	 * Setup resume button
 	 */
 	btn_resume.on('click',function(){
-		$('#stage').okaynowthis('resume',function(){
+		$('#stage').okayNowThis('resume',function(){
 			btn_pause.removeAttr('disabled');
 			btn_resume.attr('disabled','disabled');
 			log_c.prepend('resumed <br>');
-			log_z.text('.okaynowthis-playing');
+			log_z.text('.okayNowThis-playing');
 		});
     console.log('this is a log');
 	});
@@ -59,7 +59,7 @@ jQuery(document).ready(function($) {
 		// enable the pause button
 		btn_pause.removeAttr('disabled');
 		// add to logs
-		log_z.text('.okaynowthis-playing');
+		log_z.text('.okayNowThis-playing');
 
 		/**
 		 * We are declaring stage inside local scope because if the reset
@@ -68,72 +68,72 @@ jQuery(document).ready(function($) {
 		let stage = $('#stage');
 
 		// build our queue
-		stage.okaynowthis('keyframe',0,function(){
+		stage.okayNowThis('keyframe',0,function(){
 			stage.css('width','10%');
 			shape1.addClass('angle');
 			log_c.prepend('function 1: 0ms delay </br>');
 			log_x.text('.keyframe-1');
 
-		}).okaynowthis('keyframe',1000,function(){
+		}).okayNowThis('keyframe',1000,function(){
 			stage.css('width','20%');
 			log_c.prepend('function 2: 1000ms delay </br>');
 			log_x.text('.keyframe-2');
 
-		}).okaynowthis('keyframe',500,function(){
+		}).okayNowThis('keyframe',500,function(){
 			stage.css('width','30%');
 			body.addClass('rain-on');
 			log_c.prepend('function 3: 500ms delay </br>');
 			log_x.text('.keyframe-3');
 
-		}).okaynowthis('keyframe',1000,function(){
+		}).okayNowThis('keyframe',1000,function(){
 			stage.css('width','40%');
 			body.addClass('triangle-in');
 			log_c.prepend('function 4: 1000ms delay </br>');
 			log_x.text('.keyframe-4');
 
-		}).okaynowthis('keyframe',1500,function(){
+		}).okayNowThis('keyframe',1500,function(){
 			stage.css('width','50%');
 			log_c.prepend('function 5: 1500ms delay </br>');
 			log_x.text('.keyframe-5');
 
-		}).okaynowthis('keyframe',700,function(){
+		}).okayNowThis('keyframe',700,function(){
 			shape1.addClass('circle');
 			stage.css('width','60%');
 			log_c.prepend('function 6: 700ms delay </br>');
 			log_x.text('.keyframe-6');
 
-		}).okaynowthis('keyframe',1000,function(){
+		}).okayNowThis('keyframe',1000,function(){
 			stage.css('width','70%');
 			body.addClass('block-in');
 			body.addClass('triangle-out');
 			log_c.prepend('function 7: 1000ms delay </br>');
 			log_x.text('.keyframe-7');
 
-		}).okaynowthis('keyframe',600,function(){
+		}).okayNowThis('keyframe',600,function(){
 			stage.css('width','80%');
 			body.addClass('vertical-rain-on');
 			log_c.prepend('function 8: 600ms delay </br>');
 			log_x.text('.keyframe-8');
 
-		}).okaynowthis('keyframe',1000,function(){
+		}).okayNowThis('keyframe',1000,function(){
 			stage.css('width','90%');
 			shape1.removeClass('angle circle');
 			log_c.prepend('function 9: 1000ms delay </br>');
 			log_x.text('.keyframe-9');
 
-		}).okaynowthis('keyframe',700,function(){
+		}).okayNowThis('keyframe',700,function(){
 			stage.css('width','100%');
 			body.removeClass('angle circle');
 			body.addClass('block-out');
 			log_c.prepend('function 10: 700ms delay </br>');
 			log_x.text('.keyframe-10');
 
-		}).okaynowthis('keyframe',800,function(){
+		}).okayNowThis('keyframe',800,function(){
 			stage.css('width','1%');
 			log_c.prepend('function 11: 800ms delay </br>');
 			log_x.text('.keyframe-11');
 
-		}).okaynowthis('keyframe',600,function(){
+		}).okayNowThis('keyframe',600,function(){
 			log_c.prepend('function 12: 600ms delay </br>');
 			log_x.text('.keyframe-12');
 
@@ -152,7 +152,7 @@ jQuery(document).ready(function($) {
 		var wrap = $('#stage-wrap');
 		var shape1 = $('#shape-1');
 
-		stage.okaynowthis('reset',function(){
+		stage.okayNowThis('reset',function(){
 			// delete the entire staging element
 			stage.remove();
 			// add the original stage markup back in (for complex markup consider using .clone on document ready)